[Refactor] Spliting layouts related in separated file
