+++ Dev/demo.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import TextField from 'material-ui/TextField';
-import { Keyboard, RequestCloseHandler, InputHandler, extendedKeyboard } from './../src/index';
+import { Keyboard, RequestCloseHandler, InputHandler } from './../src/index';
+import { extendedKeyboard } from './../src/layouts';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import { MuiTheme } from 'material-ui/styles';
