+++ example/simple_usage/index.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import TextField from 'material-ui/TextField';
-import { Keyboard, RequestCloseHandler, InputHandler, extendedKeyboard } from 'react-material-ui-keyboard';
+import { Keyboard, RequestCloseHandler, InputHandler } from 'react-material-ui-keyboard';
+import { extendedKeyboard } from 'react-material-ui-keyboard/layouts';
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 import { MuiTheme } from 'material-ui/styles';
