+++ src/layouts.ts
@@ -1,0 +1,32 @@
+export type KeyboardLayout = Array<Array<string>>;
+
+export const alphaNumericKeyboard: KeyboardLayout = [
+    ['1',      '2',        '3', '4', '5', '6', '7', '8', '9',         '0'],
+    ['q',      'w',        'e', 'r', 't', 'y', 'u', 'i', 'o',         'p'],
+    ['a',      's',        'd', 'f', 'g', 'h', 'j', 'k', 'l', 'Backspace'],
+    ['Escape', 'CapsLock', 'z', 'x', 'c', 'v', 'b', 'n', 'm',     'Enter']
+];
+
+export const extendedKeyboard: KeyboardLayout = [
+    ['1',        '2', '3', '4', '5', '6', '7', '8', '9',         '0'],
+    ['q',        'w', 'e', 'r', 't', 'y', 'u', 'i', 'o',         'p'],
+    ['a',        's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', 'Backspace'],
+    ['CapsLock', 'z', 'x', 'c', 'v', 'b', 'n', 'm', '-',  'CapsLock'],
+    ['Escape',   '@', '',         '     ',         '.',     'Enter']
+];
+
+export const numericKeyboard: KeyboardLayout = [
+    ['Escape', '-', 'Backspace'],
+    ['7',      '8',         '9'],
+    ['4',      '5',         '6'],
+    ['1',      '2',         '3'],
+    ['0',      '.',     'Enter']
+];
+
+export function kyeboardCapsLockLayout(layout: KeyboardLayout, caps: boolean): KeyboardLayout {
+    return layout.map((row: Array<string>): Array<string> => {
+        return row.map((key: string): string => {
+            return (key.length === 1) ? (caps ? key.toUpperCase() : key.toLowerCase()) : key;
+        }); 
+    });
+}
